+++ powershell-awesome-framework/ReleaseNotes.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - `Get-PAFScriptBlockName`
 - show menu after invoke snippet script
+- Support for no catagory and function name
 
 ### Changed